Official Profiles
MakerGear profiles will be included by default in PrusaSlicer version 2.6+ so this repository is now going to be archived. Thank you for participating in the beta and for all of the feedback!

From now on, I'll be making future updates to the MakerGear PrusaSlicer profiles on the following repository:
https://github.com/Garr-R/PrusaSlicer-settings

When a future profile update gets sent to Prusa's servers you'll see an update notification within your PrusaSlicer program!